requirements.txt
# Core dependencies
yt-dlp>=2024.0.0
openai-whisper>=20231117
pysrt>=1.1.2

# Video processing
ffmpeg-python>=0.2.0

# LLM integration
ollama>=0.1.0
openai>=1.0.0
anthropic>=0.18.0

# Bilibili upload
bilibili-api-python

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Data handling
pandas>=2.0.0
pysrt>=1.1.2
